Decompose DynamicQuotaOrchestrator reconciler into dedicated dqo subpackage
Move the DynamicQuotaOrchestrator controller and helper logic into a dedicated
subpackage pkg/controller/core/dqo:
- reconciler.go: Reconciler struct, options, manager watches, and reconcile loop
- reconciler_test.go: Discovery and reconcile loop unit tests
- discovery.go: Phase 1 capacity discovery and provider aggregation logic
- distribution.go: Phase 2 subtree traversal, soft validation, and proportional quota distribution logic
- distribution_test.go: Phase 2 distribution and soft validation unit tests

Update core controller registration and integration test suite to reference
the new dqo subpackage.
